Refactor RepositoryTree from RecursiveGitTreeModel
The aim here is to separate a reusable model for the GitHub API
(RecursiveGitTreeModel) from the find_notebooks method which _is_ domain
specific and now belongs on the RepositoryTree domain model.
