+++ tests/domain/githubcheckout_test.py
@@ -9,49 +9,13 @@ import pytest
 import respx
 from gidgethub.httpx import GitHubAPI
 from httpx import Response
-from safir.github.models import GitHubBlobModel
 
 from timessquare.domain.githubcheckout import (
     GitHubRepositoryCheckout,
-    GitTreeMode,
-    RecursiveGitTreeModel,
+    RepositoryTree,
 )
+from timessquare.storage.github.apimodels import RecursiveGitTreeModel
 from timessquare.storage.github.settingsfiles import RepositorySettingsFile
-
-
-def test_recursive_git_tree_model_rsp_broadcast() -> None:
-    """Test that an object returned by the GitHub Git Tree API with
-    recursive=1 can be parsed by RecursiveGitTreeModel.
-    """
-    json_path = Path(__file__).parent.joinpath(
-        "../data/rsp_broadcast/recursive_tree.json"
-    )
-    repo_tree = RecursiveGitTreeModel.model_validate_json(
-        json_path.read_text()
-    )
-    assert repo_tree.sha == "46372dfa5a432026d68d262899755ef0333ef8c0"
-    assert repo_tree.truncated is False
-    assert len(repo_tree.tree) == 14
-
-    for tree_item in repo_tree.tree:
-        if tree_item.path == "README.md":
-            assert tree_item.mode == GitTreeMode.file
-            assert tree_item.sha == "8e977bc4a1503adb11e3fe06e0ddcf759ad59a91"
-
-
-def test_git_blob_model_rsp_broadcast() -> None:
-    """Test that a git blob, as retrieved from the GitTreeItem.url attribute
-    can be parsed by GitHubBlobModel.
-    """
-    json_path = Path(__file__).parent.joinpath(
-        "../data/rsp_broadcast/readme_blob.json"
-    )
-    blob = GitHubBlobModel.model_validate_json(json_path.read_text())
-    assert blob.sha == "8e977bc4a1503adb11e3fe06e0ddcf759ad59a91"
-    assert blob.encoding == "base64"
-    assert blob.decode().startswith(
-        "# Broadcast messages for the Rubin Science Platform"
-    )
 
 
 @pytest.mark.asyncio
@@ -88,17 +52,21 @@ async def test_repository_git_tree(
         "46372dfa5a432026d68d262899755ef0333ef8c0?recursive=1"
     ).mock(return_value=Response(200, json=json.loads(json_path.read_text())))
     repo_tree = await repo.get_git_tree(github_client)
-    assert repo_tree.sha == "46372dfa5a432026d68d262899755ef0333ef8c0"
+    assert (
+        repo_tree.github_tree.sha == "46372dfa5a432026d68d262899755ef0333ef8c0"
+    )
 
 
 @pytest.mark.asyncio
 async def test_recursive_git_tree_find_notebooks() -> None:
-    """Test RecursiveGitTreeModel using the times-square-time dataset."""
+    """Test RepositoryTree using the times-square-time dataset."""
     json_path = Path(__file__).parent.joinpath(
         "../data/times-square-demo/recursive_tree.json"
     )
-    repo_tree = RecursiveGitTreeModel.model_validate_json(
-        json_path.read_text()
+    repo_tree = RepositoryTree(
+        github_tree=RecursiveGitTreeModel.model_validate_json(
+            json_path.read_text()
+        )
     )
     settings = RepositorySettingsFile(ignore=[])
     notebook_refs = list(repo_tree.find_notebooks(settings))
